third_party/WebKit/LayoutTests/fast/overflow/resize-inherit.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
.overflow {
    overflow: scroll;
    width: 200px;
    height: 200px;
}
.resize {
    resize: both;
}
.resize-inherit {
    resize: inherit;
}
</style>
<div class="overflow resize">
    <div class="overflow no-resize"></div>
    <div class="overflow resize-inherit"></div>
</div>
<script>
description("CSS resize property is inherited:no");

shouldBe("getComputedStyle(document.querySelector('.overflow.resize')).resize", "'both'");
shouldBe("getComputedStyle(document.querySelector('.overflow.no-resize')).resize", "'none'");
shouldBe("getComputedStyle(document.querySelector('.overflow.resize-inherit')).resize", "'both'");
</script>
