third_party/blink/web_tests/fast/overflow/resize-inherit-expected.txt

CSS resize property is inherited:no

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.querySelector('.overflow.resize')).resize is 'both'
PASS getComputedStyle(document.querySelector('.overflow.no-resize')).resize is 'none'
PASS getComputedStyle(document.querySelector('.overflow.resize-inherit')).resize is 'both'
PASS successfullyParsed is true

TEST COMPLETE
